+++ testapps/GHIssues/src/main/java/ghissues/gh1078/Gh1078_Service.java
@@ -4,7 +4,7 @@ import act.controller.annotation.UrlContext;
 import org.osgl.mvc.annotation.GetAction;
 
 @UrlContext("1078")
-public class Service extends Base {
+public class Gh1078_Service extends Gh1078_Base {
 
     @GetAction
     public String test() {